feat: move common methods for resuse
